test(logic): split resource_extractor part2 tests for file-size gate
Split the oversized resource_extractor_part2_test.dart into part1 and
part2 files so each stays under the 400-line logic test gate (Refs #2288).
